made weather deltas doubles

+++ config.c
@@ -23,9 +23,9 @@ const int weatherRecordingInterval = 300;
 const int weatherUpdateInterval = 95;
 const int energyUpateInterval = 116;
 
-const int temperatureDeltaForChange = 1;
-const int pressureDeltaForChange = 1;
-const int humidityDeltaForChange = 1;
+const double temperatureDeltaForChange = 1;
+const double pressureDeltaForChange = 1;
+const double humidityDeltaForChange = 1;
 
 const int backlightTimeout = 300; //how long the backlight stays lit after we start up or a button is pushed.  the backlight uses 100 times more power than the rest of the LCD module
